Introduce `DatabaseHelper` class and refactor `SqliteStorage`
- Created `DatabaseHelper` class to handle db initialization, table
  creation
- Refactor `SqliteStorage` to use `DatabaseHelper`
