+++ packages/sqlite_storage/lib/src/sqlite_storage.dart
@@ -1,45 +1,20 @@
 import 'package:domain_models/domain_models.dart';
-import 'package:path/path.dart';
 import 'package:sqflite/sqflite.dart';
+import 'package:sqlite_storage/src/database_helper.dart';
 
 import 'exceptions.dart';
 
 class SqliteStorage {
-  const SqliteStorage();
+  final DatabaseHelper _dbHelper = DatabaseHelper();
 
-  static Database? _database;
+  Database? _database;
 
   Future<void> init() async {
-    final String path = join(
-      await getDatabasesPath(),
-      'footprint.db',
-    );
-
-    try {
-      _database = await openDatabase(
-        path,
-        version: 1,
-        onCreate: _createTables,
-      );
-    } on DatabaseException catch (_) {
-      throw UnableCreateDatabaseException();
-    }
+    _database = await _dbHelper.database;
   }
 
   Future<void> close() async {
-    if (_database != null) {
-      await _database!.close();
-    }
-  }
-
-  Future<void> _createTables(Database db, int version) async {
-    try {
-      await _Routes.createTable(db);
-      await _RoutePoints.createTable(db);
-      await _GeocodingCache.createTable(db);
-    } on DatabaseException catch (_) {
-      throw UnableCreateTableException();
-    }
+    await _dbHelper.close();
   }
 
   Future<int> createRoute(RoutePoint routePoint) async {
@@ -49,7 +24,7 @@ class SqliteStorage {
       }
 
       final int routeId = await _database!.insert(
-        _Routes.tableName,
+        Routes.tableName,
         <String, dynamic>{
           'start_time': DateTime.now().toIso8601String(),
           'status': 'active',
@@ -57,7 +32,7 @@ class SqliteStorage {
       );
 
       final int routePointsId = await _database!.insert(
-        _RoutePoints.tableName,
+        RoutePoints.tableName,
         <String, dynamic>{
           'route_id': routeId,
           'latitude': routePoint.latitude,
@@ -72,20 +47,16 @@ class SqliteStorage {
     }
   }
 
-  Future<int> upsertRoutePointOfRouteById(
-    int routeId,
-    RoutePoint routePoint,
-  ) async {
+  Future<int> insertRoutePoint(int routeId, RoutePoint routePoint) async {
     try {
       if (_database == null) {
         await init();
       }
 
-      return await _database!.update(
-        _RoutePoints.tableName,
-        where: 'route_id = ?',
-        whereArgs: [routeId],
+      return await _database!.insert(
+        RoutePoints.tableName,
         <String, dynamic>{
+          'route_id': routeId,
           'latitude': routePoint.latitude,
           'longitude': routePoint.longitude,
           'address': routePoint.address,
@@ -97,14 +68,14 @@ class SqliteStorage {
     }
   }
 
-  Future<Route> getPointsOfRouteById(int routeId) async {
+  Future<Route> getAllRoutePoints(int routeId) async {
     try {
       if (_database == null) {
         await init();
       }
 
       final List<Map<String, dynamic>> result = await _database!.query(
-        _RoutePoints.tableName,
+        RoutePoints.tableName,
         where: 'route_id = ?',
         whereArgs: [routeId],
         orderBy: 'id ASC',
@@ -116,7 +87,7 @@ class SqliteStorage {
       );
 
       final route = await _database!.query(
-        _Routes.tableName,
+        Routes.tableName,
         where: 'id = ?',
         whereArgs: [routeId],
       );
@@ -139,7 +110,7 @@ class SqliteStorage {
     }
   }
 
-  Future<void> deleteRoute(int routeId) async {
+  Future<void> deleteRouteById(int routeId) async {
     try {
       if (_database == null) {
         await init();
@@ -155,13 +126,13 @@ class SqliteStorage {
       // });
       await _database!.transaction((Transaction txn) async {
         await txn.delete(
-          _RoutePoints.tableName,
+          RoutePoints.tableName,
           where: 'route_id = ?',
           whereArgs: [routeId],
         );
 
         await txn.delete(
-          _Routes.tableName,
+          Routes.tableName,
           where: 'id = ?',
           whereArgs: [routeId],
         );
@@ -178,7 +149,7 @@ class SqliteStorage {
       }
 
       return await _database!.update(
-        _Routes.tableName,
+        Routes.tableName,
         where: 'id = ?',
         whereArgs: [routeId],
         <String, dynamic>{
@@ -188,58 +159,5 @@ class SqliteStorage {
     } on DatabaseException catch (_) {
       throw UnableUpdateDatabaseException();
     }
-  }
-}
-
-class _Routes {
-  static const String tableName = 'routes';
-
-  static Future<void> createTable(Database db) async {
-    await db.execute('''
-    CREATE TABLE $tableName (
-      id            INTEGER PRIMARY KEY AUTOINCREMENT,
-      start_point   TEXT,
-      end_point     TEXT,
-      start_time    TEXT,
-      end_time      TEXT,
-      distance      REAL,
-      average_speed REAL,
-      status        TEXT CHECK(status IN ('active', 'completed')), 
-    );
-    ''');
-  }
-}
-
-class _RoutePoints {
-  static const String tableName = 'route_points';
-
-  static Future<void> createTable(Database db) async {
-    await db.execute('''
-    CREATE TABLE $tableName (
-      id              INTEGER PRIMARY KEY AUTOINCREMENT,
-      route_id        INTEGER NOT NULL,
-      latitude        REAL NOT NULL,
-      longitude       REAL NOT NULL,  
-      address         TEXT,
-      timestamp       TEXT NOT NULL,
-      FOREIGN KEY(route_id) REFERENCES routes(id)
-    );
-    ''');
-  }
-}
-
-class _GeocodingCache {
-  static const String tableName = 'geocoding_cache';
-
-  static Future<void> createTable(Database db) async {
-    await db.execute('''
-    CREATE TABLE $tableName (
-      id        INTEGER PRIMARY KEY AUTOINCREMENT,
-      latitude  REAL NOT NULL,
-      longitude REAL NOT NULL,
-      usage_frequency INTEGER DEFAULT 0,
-      timestamp TEXT NOT NULL,
-    );
-    ''');
   }
 }
